fastapi/requirements.txt
google-generativeai
graphdatascience
retry==0.9.2
langchain>=0.0.216
langchain_community
numpy
pandas
plotly==5.15.0
altair<5
python-dotenv==1.0.0
langchain-google-genai 
pillow
starlette==0.37.2
re
youtube_transcript_api==0.6.2
urllib3==2.2.2
wikipedia==1.4.0
SpeechRecognition==3.10.4
fastapi==0.111.0
fastapi-cli==0.0.4